# docs/04-API/openapi.yaml
openapi: 3.0.3
info:
  title: 教务管理系统 API
  description: |
    培训机构**纯员工端 CRM** 后端 API（Express 4 + node:sqlite）。
    仅 `admin` / `teacher` 两种员工角色；学生、家长无账号、不登录。

    本文件是机器可读规范，人类可读的完整说明见 `docs/04-API/API.md`。
    覆盖范围：认证 + 四个 P0 模块（考勤签到 / 学生 / 成绩 / 招生线索）+ 通知 + AI 分析端点 + 课次实体与任课关系（v21）。
    统一响应包裹：成功 `{ success: true, data }`；失败 `{ success: false, message }`。
    `/api/analytics/*` 使用独立的「AI 分析信封」。
    v21（2026-09-23）：新增 `/api/sessions/*`（课次）、`/api/teaching-assignments`（任课关系）、`/api/period-times`（节次时间）；
    `POST /api/attendance/batch`、`GET /api/attendance`、`POST /api/growth/class-eval` 增加可选 `session_id`。
  version: 1.0.0
  contact:
    name: 教务管理系统
servers:
  - url: http://localhost:3000
    description: 本地 / Docker 后端
...
security:
  - bearerAuth: []
paths:
...
  /api/teaching-assignments:
    get:
      tags: [TeachingAssignments]
      summary: 任课关系列表（教师仅本班）
      security: [{bearerAuth: []}]
      parameters:
        - { in: query, name: class_id, schema: { type: integer } }
        - { in: query, name: teacher_id, schema: { type: integer } }
        - { in: query, name: term_id, schema: { type: integer } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, default: 10 } }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      list: { type: array, items: { $ref: "#/components/schemas/TeachingAssignment" } }
                      total: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [TeachingAssignments]
      summary: 新增任课关系（仅 admin）
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeachingAssignmentInput" }
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties: { id: { type: integer } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/teaching-assignments/{id}:
    put:
      tags: [TeachingAssignments]
      summary: 修改任课关系（仅 admin）
      security: [{bearerAuth: []}]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeachingAssignmentInput" }
      responses:
        "200": { description: 成功, content: { application/json: { schema: { $ref: "#/components/schemas/SuccessEnvelope" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: 任课关系不存在, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
    delete:
      tags: [TeachingAssignments]
      summary: 删除任课关系（仅 admin）
      security: [{bearerAuth: []}]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      responses:
        "200": { description: 成功, content: { application/json: { schema: { $ref: "#/components/schemas/SuccessEnvelope" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { description: 任课关系不存在, content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } } }
  /api/period-times:
    get:
      tags: [PeriodTimes]
      summary: 节次时间表（1–8 节起止时间）
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { type: array, items: { $ref: "#/components/schemas/PeriodTime" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    put:
      tags: [PeriodTimes]
      summary: 批量保存节次时间（仅 admin，仅 1–8 节）
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required: [period]
                    properties:
                      period: { type: integer, minimum: 1, maximum: 8 }
                      start_time: { type: string, description: HH:mm 或空串 }
                      end_time: { type: string, description: HH:mm 或空串 }
                      label: { type: string, maxLength: 20 }
      responses:
        "200": { description: 成功, content: { application/json: { schema: { $ref: "#/components/schemas/SuccessEnvelope" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: 参数错误或业务规则拒绝
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    Unauthorized:
      description: 未登录 / Token 失效 / 凭证已被吊销（登出、改密、改角色后 token_version 不再匹配）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    Forbidden:
      description: 无权限（教师越权 / 非 admin）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success: { type: boolean, example: true }
        data: { type: object }
    ErrorEnvelope:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, example: 该学员已有报班记录，不能删除 }
...
    TeachingAssignment:
      type: object
      properties:
        id: { type: integer }
        class_id: { type: integer }
        class_name: { type: string }
        course_id: { type: integer }
        course_name: { type: string }
        teacher_id: { type: integer, nullable: true }
        teacher_name: { type: string, nullable: true }
        term_id: { type: integer, nullable: true }
        term_name: { type: string, nullable: true }
    TeachingAssignmentInput:
      type: object
      required: [class_id, course_id]
      properties:
        class_id: { type: integer }
        course_id: { type: integer }
        teacher_id: { type: integer, nullable: true, description: 可空=未定教师 }
        term_id: { type: integer, nullable: true, description: 可空=长期有效 }
    PeriodTime:
      type: object
      properties:
        period: { type: integer, minimum: 1, maximum: 8 }
        start_time: { type: string }
        end_time: { type: string }
        label: { type: string }